refactor: move bulk_grade_submissions to assignments.py
Relocate bulk grading from rubrics.py to assignments.py where it
belongs (supports both rubric and point-based grading). Import
build_rubric_assessment_form_data from rubrics as shared utility.

Part of rubric tool rationalization (11 → 6 tools).
