+++ src/canvas_mcp/tools/assignments.py
@@ -1,7 +1,9 @@
 """Assignment-related MCP tools for Canvas API."""
 
+import asyncio
 import datetime
 from statistics import StatisticsError, mean, median, stdev
+from typing import Any
 
 from mcp.server.fastmcp import FastMCP
 from mcp.types import ToolAnnotations
@@ -12,6 +14,7 @@ from ..core.client import fetch_all_paginated_results, make_canvas_request
 from ..core.dates import format_date, parse_date
 from ..core.logging import log_error
 from ..core.validation import validate_params
+from .rubrics import build_rubric_assessment_form_data
 
 
 def register_shared_assignment_tools(mcp: FastMCP):
@@ -942,3 +945,214 @@ def register_educator_assignment_tools(mcp: FastMCP):
             result += f"  URL: {html_url}\n"
 
         return result
+
+    @mcp.tool()
+    @validate_params
+    async def bulk_grade_submissions(
+        course_identifier: str | int,
+        assignment_id: str | int,
+        grades: dict[str, Any],
+        dry_run: bool = False,
+        max_concurrent: int = 5,
+        rate_limit_delay: float = 1.0
+    ) -> str:
+        """Grade multiple submissions efficiently with concurrent processing.
+
+        Supports both rubric-based and simple point-based grading in batches.
+
+        Args:
+            course_identifier: Course code or Canvas ID
+            assignment_id: Canvas assignment ID
+            grades: Dict mapping user_id to {rubric_assessment?, grade?, comment?}
+            dry_run: If True, validate without submitting (default: False)
+            max_concurrent: Max concurrent grading operations (default: 5)
+            rate_limit_delay: Delay between batches in seconds (default: 1.0)
+        """
+        course_id = await get_course_id(course_identifier)
+        assignment_id_str = str(assignment_id)
+
+        # Validate that we have grades to process
+        if not grades:
+            return "Error: No grades provided. The grades dictionary is empty."
+
+        # Check if rubric is configured for grading (if using rubric assessments)
+        has_rubric_grades = any(
+            "rubric_assessment" in grade_info
+            for grade_info in grades.values()
+        )
+
+        if has_rubric_grades:
+            assignment_check = await make_canvas_request(
+                "get",
+                f"/courses/{course_id}/assignments/{assignment_id_str}",
+                params={"include[]": ["rubric_settings"]}
+            )
+
+            if "error" not in assignment_check:
+                use_rubric_for_grading = assignment_check.get("use_rubric_for_grading", False)
+                if not use_rubric_for_grading and not dry_run:
+                    return (
+                        "⚠️  ERROR: Rubric is not configured for grading!\n\n"
+                        "The rubric exists but 'use_for_grading' is set to FALSE.\n"
+                        "Grades will NOT be saved to the gradebook.\n\n"
+                        "To fix this:\n"
+                        "1. Use get_rubric to verify rubric settings\n"
+                        "2. Use associate_rubric with use_for_grading=True\n"
+                        "3. Or set dry_run=True to test without submitting\n"
+                    )
+
+        # Statistics tracking
+        stats = {
+            "total": len(grades),
+            "graded": 0,
+            "failed": 0
+        }
+        failed_results = []
+
+        async def grade_single_submission(user_id: str, grade_info: dict[str, Any]):
+            """Grade a single submission."""
+            try:
+                if dry_run:
+                    # In dry run mode, just validate the data
+                    if "rubric_assessment" in grade_info:
+                        total_points = sum(
+                            criterion.get("points", 0)
+                            for criterion in grade_info["rubric_assessment"].values()
+                        )
+                        return {
+                            "status": "success",
+                            "user_id": user_id,
+                            "message": f"DRY RUN: Would grade with {total_points} rubric points"
+                        }
+                    elif "grade" in grade_info:
+                        return {
+                            "status": "success",
+                            "user_id": user_id,
+                            "message": f"DRY RUN: Would grade with {grade_info['grade']} points"
+                        }
+                    else:
+                        return {
+                            "status": "failed",
+                            "user_id": user_id,
+                            "error": "No rubric_assessment or grade provided"
+                        }
+
+                # Build form data based on grading type
+                form_data = {}
+
+                if "rubric_assessment" in grade_info and grade_info["rubric_assessment"]:
+                    # Rubric-based grading
+                    form_data = build_rubric_assessment_form_data(
+                        grade_info["rubric_assessment"],
+                        grade_info.get("comment")
+                    )
+                elif "grade" in grade_info:
+                    # Simple grading
+                    form_data["submission[posted_grade]"] = str(grade_info["grade"])
+                    if "comment" in grade_info:
+                        form_data["comment[text_comment]"] = grade_info["comment"]
+                else:
+                    return {
+                        "status": "failed",
+                        "user_id": user_id,
+                        "error": "Must provide either rubric_assessment or grade"
+                    }
+
+                # Submit the grade
+                response = await make_canvas_request(
+                    "put",
+                    f"/courses/{course_id}/assignments/{assignment_id_str}/submissions/{user_id}",
+                    data=form_data,
+                    use_form_data=True
+                )
+
+                if "error" in response:
+                    return {
+                        "status": "failed",
+                        "user_id": user_id,
+                        "error": response["error"]
+                    }
+
+                return {
+                    "status": "success",
+                    "user_id": user_id,
+                    "grade": response.get("grade", "N/A")
+                }
+
+            except Exception as e:
+                return {
+                    "status": "failed",
+                    "user_id": user_id,
+                    "error": str(e)
+                }
+
+        # Process in batches
+        user_ids = list(grades.keys())
+        total_batches = (len(user_ids) + max_concurrent - 1) // max_concurrent
+
+        result_lines = []
+        result_lines.append(f"{'=' * 60}")
+        result_lines.append(f"Bulk Grading {'(DRY RUN) ' if dry_run else ''}for Assignment {assignment_id}")
+        result_lines.append(f"{'=' * 60}")
+        result_lines.append(f"Course: {await get_course_code(course_id) or course_identifier}")
+        result_lines.append(f"Total submissions to grade: {stats['total']}")
+        result_lines.append(f"Concurrent processing: {max_concurrent} per batch")
+        result_lines.append(f"Total batches: {total_batches}\n")
+
+        for i in range(0, len(user_ids), max_concurrent):
+            batch = user_ids[i:i + max_concurrent]
+            batch_num = (i // max_concurrent) + 1
+
+            result_lines.append(f"Processing batch {batch_num}/{total_batches} ({len(batch)} submissions)...")
+
+            # Process batch concurrently
+            tasks = [
+                grade_single_submission(user_id, grades[user_id])
+                for user_id in batch
+            ]
+            results = await asyncio.gather(*tasks, return_exceptions=True)
+
+            # Update statistics
+            for result in results:
+                if isinstance(result, Exception):
+                    stats["failed"] += 1
+                    failed_results.append({
+                        "user_id": "unknown",
+                        "error": str(result)
+                    })
+                elif result["status"] == "success":
+                    stats["graded"] += 1
+                    result_lines.append(f"  ✓ User {result['user_id']}: {result.get('message', 'Graded')}")
+                else:
+                    stats["failed"] += 1
+                    failed_results.append({
+                        "user_id": result["user_id"],
+                        "error": result["error"]
+                    })
+                    result_lines.append(f"  ✗ User {result['user_id']}: {result['error']}")
+
+            # Rate limit between batches (except after last batch)
+            if i + max_concurrent < len(user_ids):
+                result_lines.append(f"  Waiting {rate_limit_delay}s before next batch...\n")
+                await asyncio.sleep(rate_limit_delay)
+
+        # Summary
+        result_lines.append(f"\n{'=' * 60}")
+        result_lines.append(f"Bulk Grading {'(DRY RUN) ' if dry_run else ''}Complete!")
+        result_lines.append(f"{'=' * 60}")
+        result_lines.append(f"Total:   {stats['total']}")
+        result_lines.append(f"Graded:  {stats['graded']}")
+        result_lines.append(f"Failed:  {stats['failed']}")
+
+        if failed_results:
+            result_lines.append("\nFailed Submissions:")
+            for failure in failed_results[:10]:  # Show first 10 failures
+                result_lines.append(f"  User {failure['user_id']}: {failure['error']}")
+            if len(failed_results) > 10:
+                result_lines.append(f"  ... and {len(failed_results) - 10} more failures")
+
+        if dry_run:
+            result_lines.append("\n⚠️  DRY RUN MODE: No grades were actually submitted")
+            result_lines.append("Set dry_run=false to apply grades")
+
+        return "\n".join(result_lines)
